fix: resolve 13 Minor issues from Phase 2 code review
Performance:
- M1: Shared AsyncOpenAI client singleton instead of per-request instantiation
- M7: Add LIMIT 200 to evidence and timeline queries

Security & correctness:
- M3: Change evidence upload params from Query to Form for multipart compatibility
- M4: Read file content and use len() instead of unreliable file.size
- M5: Lower page_size max from 100 to 50 to prevent expensive queries
- M6: Add pagination to favorites list endpoint
- M8: Move OpenAI client creation outside SSE generator
- M11: Change file_size from BigInteger to Integer (sufficient for 10MB cap)

Code quality:
- M2: Fix model name casing in docstrings (glm-5.1, glm-5-turbo)
- M9: Remove redundant pop/push of template_id in document creation
- M13: Add truncation note when articles exceed 20 in law interpretation
- M15: Fix list_templates to explicitly handle None tenant_id

+++ backend/app/models/case.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import uuid
 from datetime import date, datetime
 
-from sqlalchemy import ARRAY, BigInteger, Date, ForeignKey, Integer, Numeric, String, Text
+from sqlalchemy import ARRAY, Date, ForeignKey, Integer, Numeric, String, Text
 from sqlalchemy.dialects.postgresql import JSONB, UUID
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
@@ -41,7 +41,7 @@ class Evidence(Base, BaseMixin):
     title: Mapped[str] = mapped_column(String(200))
     evidence_type: Mapped[str] = mapped_column(String(50))
     file_url: Mapped[str | None] = mapped_column(String(500))
-    file_size: Mapped[int | None] = mapped_column(BigInteger)
+    file_size: Mapped[int | None] = mapped_column(Integer)
     file_type: Mapped[str | None] = mapped_column(String(50))
     description: Mapped[str | None] = mapped_column(Text)
     sort_order: Mapped[int] = mapped_column(Integer, default=0)
